requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
python-rtmidi==1.4.9
toml==0.10.2
